chore: refactor session handler (#224)
